+++ src/especialista.cpp
@@ -30,7 +30,7 @@ Especialista::Especialista(){
     this->start_ociosity = 0;
 }
 /* SET METHODS */
-void Especialista::set_id(u_int8_t id){
+void Especialista::set_id(u_int16_t id){
     this->id = id;
 }
 void Especialista::set_situation(state_esp situation){
@@ -43,7 +43,7 @@ void Especialista::set_start_ociosity(u_int32_t start_ociosity){
     this->start_ociosity = start_ociosity;
 }
 /* GET METHODS */
-u_int8_t  Especialista::get_id(){
+u_int16_t  Especialista::get_id(){
     return this->id;
 }
 state_esp Especialista::get_situation(){
